General update. Working towards getting new driver interface, This patch starts the work of getting rid of Mach-inspired components in the kernel. We want to be closer to the ReactOS kernel (in terms of layout and general APi structure) in an effort to make drivers for various peripherals much easier to port over (as ReactOS-like operating systems have much more hardware compatible for it than Linux or BSD).
As such, we're going to deprecate (and eventually remove) the old ddk,
and some other old components, eventually completely redesigning the
kernel so we get the 'theoretical fun stuff' to connect to the actual,
working, bootable part in the HALs.

We're going to shift from using Kernel_Main() as the entry point to
KiSystemStartup(), and actually adopt the function naming convention
initially proposed anyway.